move JSExports to dedicated file, make UpdateRootComponents to use WebAssemblyCallQueue
